+++ src/main/java/com/plate/silverplate/battle/domain/entity/Battle.java
@@ -1,4 +1,4 @@
-package com.plate.silverplate.domain;
+package com.plate.silverplate.battle.domain.entity;
 
 import com.plate.silverplate.common.entity.BaseTimeEntity;
 import jakarta.persistence.*;
