+++ src/main/java/com/plate/silverplate/meal/domain/entity/Meal.java
@@ -1,4 +1,4 @@
-package com.plate.silverplate.domain;
+package com.plate.silverplate.meal.domain.entity;
 
 import com.plate.silverplate.common.entity.BaseTimeEntity;
 import jakarta.persistence.*;
