+++ src/main/java/com/plate/silverplate/nutritionFact/domain/entity/NutritionFact.java
@@ -1,4 +1,4 @@
-package com.plate.silverplate.domain;
+package com.plate.silverplate.nutritionFact.domain.entity;
 
 import com.plate.silverplate.common.entity.BaseTimeEntity;
 import jakarta.persistence.*;
