+++ src/main/java/com/plate/silverplate/user/domain/entity/User.java
@@ -1,12 +1,11 @@
-package com.plate.silverplate.domain;
+package com.plate.silverplate.user.domain.entity;
 
 import com.plate.silverplate.common.entity.BaseTimeEntity;
+import com.plate.silverplate.userPhysical.domain.entity.UserPhysical;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.*;
-
-import java.time.Instant;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
